Add /next_revision prompt
the prompt allows to summarize the feedback in files.
The use-case is to later feed git notes with them.

git-notes can be directly invoked by the model, but write
instructions to the repository has been intentionally avoided,
for the time being.

While at it, create my_status_prompt, reformat and move the instructions in a
variable for that case as well.

Signed-off-by: Paolo Valerio <pvalerio@redhat.com>